Removed old icon tests
