+++ mapstory/tests/js/composer.po.js
@@ -7,7 +7,7 @@
 require('./waitReady.js');
 
 var composer = function() {
-	this.startComposingLink = element(by.css("#start-composing"));
+	this.startComposingLink = element(by.css('#start-composing'));
 
 	this.get = function() {
 		browser.get('http://192.168.56.151');
