+++ mapstory/tests/js/conf.js
@@ -1,5 +1,16 @@
 'use strict';
 
+/**
+ * End to end test Config
+ * ======================
+ *
+ * Notes
+ * ----
+ * - Will automatically include all tests inside files named `*.spec.js`
+ * - Sometimes tests break in chrome when you have another instance of the browser open and logged into mapstory.
+ *
+ */
+
 exports.config = {
 	framework: 'jasmine',
 	seleniumAddress: 'http://localhost:4444/wd/hub',
